requirements.txt
nose==1.3.0
flask~=1.1.2
python-decouple==3.4
aiohttp~=3.7.4
redis~=3.5.3
requests~=2.25.1